srcds1 "start" command micro-fix
Fixed bug, when user might to execute "/etc/init.d/srcds1 start" 2 times or more. This led to the fact that the process was duplicated and thus it was possible to run several servers with the same process name (for example, "L4D2 Dedicated Server 1"), but with different ports. This fix, when you re-enter the command "/etc/init.d/srcds1 start", will give a message that the process has already started and will suspend further execution of the command.